Move `StructResponse` under `marimo._server.responses`
